[CODE_REVIEW] Move RideAxis to _utils since it's not an official component
