+++ src/index.tsx
@@ -52,7 +52,6 @@ export * from './pushInfo'
 export * from './qrCard'
 export * from './rating'
 export * from './review'
-export * from './rideAxis'
 export * from './searchRecap'
 export * from './selectField'
 export * from './snackbar'
@@ -74,6 +73,7 @@ export * from './tripCard'
 export * from './uneditableTextField'
 export * from './why'
 
+export * from './_utils/rideAxis'
 export * from './_utils/branding'
 export * from './_utils/onChange'
 export * from './_utils/place'
