+++ stories/index.tsx
@@ -94,6 +94,7 @@ import '../src/title/story'
 import '../src/toggleButton/story'
 import '../src/topBar/story'
 import '../src/tripCard/story'
-import '../src/rideAxis/story'
 import '../src/uneditableTextField/story'
 import '../src/why/story'
+// INTERNAL
+import '../src/_utils/rideAxis/story'
